@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <UserSecretsId>ackablechatroom</UserSecretsId>
     <RootNamespace>Microsoft.Azure.SignalR.Samples.AckableChatRoom</RootNamespace>
   </PropertyGroup>
@@ -10,7 +10,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.2.*" />
-    <PackageReference Include="Microsoft.Azure.SignalR" Version="1.0.*" />
+    <PackageReference Include="Microsoft.Azure.SignalR" Version="1.*" />
   </ItemGroup>
 </Project>
