@@ -1,9 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <UserSecretsId>flightmap</UserSecretsId>
     <RootNamespace>Microsoft.Azure.SignalR.Samples.FlightMap</RootNamespace>
-    <UserSecretsId>flightMap</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
