@@ -25,7 +25,7 @@ namespace Microsoft.Azure.SignalR.Samples.AckableChatRoom
                     {
                         //  This is a tircky way to associate user name with connection for sample purpose.
                         //  For PROD, we suggest to use authentication and authorization, see here:
-                        //  https://docs.microsoft.com/en-us/aspnet/core/signalr/authn-and-authz?view=aspnetcore-2.2
+                        //  https://docs.microsoft.com/en-us/aspnet/core/signalr/authn-and-authz
                         options.ClaimsProvider = context => new[]
                         {
                             new Claim(ClaimTypes.NameIdentifier, context.Request.Query["username"])
@@ -37,11 +37,11 @@ namespace Microsoft.Azure.SignalR.Samples.AckableChatRoom
 
         public void Configure(IApplicationBuilder app)
         {
-            app.UseMvc();
+            app.UseRouting();
             app.UseFileServer();
-            app.UseAzureSignalR(routes =>
+            app.UseEndpoints(e =>
             {
-                routes.MapHub<AckableChatSampleHub>("/chat");
+                e.MapHub<AckableChatSampleHub>("/chat");
             }
             );
         }
