@@ -1,20 +1,9 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+﻿var builder = WebApplication.CreateBuilder(args);
+builder.Services.AddSignalR().AddAzureSignalR();
+var app = builder.Build();
 
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
-
-namespace Microsoft.Azure.SignalR.Samples.ChatRoom
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateWebHostBuilder(args).Build().Run();
-        }
-
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>();
-    }
-}
+app.UseDefaultFiles();
+app.UseStaticFiles();
+app.UseRouting();
+app.MapHub<ChatSampleHub>("/chat");
+app.Run();
